Refactor whole project by api modules
