--- NodeExceptions.py
@@ -1,18 +1,0 @@
-class NodeNotConnectedException(Exception):
-    pass
-
-
-class AddressInvalidException(Exception):
-    pass
-
-
-class NodeSyncException(Exception):
-    pass
-
-
-class TransferInvalidException(Exception):
-    pass
-
-
-class ApprovalInvalidException(Exception):
-    pass
