+++ account/DB.py
@@ -1,13 +1,12 @@
 import os
-from sys import flags
+import json
 
 from sqlalchemy import create_engine
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
-import json
 
 BASE_DIR = os.path.dirname(os.path.abspath(__file__))
-SECRET_FILE = os.path.join(BASE_DIR, 'secret.json')
+SECRET_FILE = os.path.join(BASE_DIR, '../secret.json')
 secret = json.loads(open(SECRET_FILE).read())
 DB = secret["DB"]
