--- models.py
@@ -1,13 +1,0 @@
-from sqlalchemy import Boolean, Column, ForeignKey, Integer, String
-from sqlalchemy.orm import relationship
-
-from DB import Base
-
-
-class User(Base):
-    __tablename__ = "user"
-
-    id = Column(Integer, primary_key=True, nullable=False)
-    username = Column(String, nullable=False)
-    password = Column(String, nullable=False)
-    nickname = Column(String, nullable=False)
